+++ frictionless/actions/__init__.py
@@ -1,4 +1,6 @@
 from .describe import describe
+from .extract import extract
 
 # TODO: remove these legacy imports in v5
 from .describe import describe_dialect, describe_resource, describe_package, describe_schema
+from .extract import extract_resource, extract_package
