--- frictionless/extract/__init__.py
@@ -1,3 +1,0 @@
-from .main import extract
-from .package import extract_package
-from .resource import extract_resource
