+++ frictionless/plugins/server/server.py
@@ -1,6 +1,5 @@
 import multiprocessing
-from ...actions import describe
-from ...extract import extract
+from ...actions import describe, extract
 from ...validate import validate
 from ...transform import transform
 from ...server import Server
