+++ tests/actions/extract/test_resource.py
